# An Unobtrusive LRU for the best time cache I've used for go