I needed a time AND LRU cache. By far, this is the best time cache I've found for go. As such, I've injected LRU capabilities that simply:
- ensures that when an item is added, if the number of unexpired items >= maximum number of items specified, the oldest (based on ACCESS) is expired
- when an item is "gotten", or "set", the access time is updated

The existing janitor system is unchanged, and used to actually delete the items.

The only change from a consumer standpoint, is one additional parameter at the end of a "New" call, that specifies the maximum number of items to allow, or 0, if you want to disable the LRU completely.

I updated the "test" file so that it would run as-is with LRU disabled. I used my own test harness that isn't horribly portable. I'll work on updating yours with LRU tests in my Copious Free Time (patent pending) if you're interested in accepting this pull.

This is 1/2 of what #5 is asking for.