@pmylund 2015-03-03
Thanks for the pull request.

This is a big change, and my main concern is that it breaks API compatibility. If this is to be merged, it needs to be changed so that the LRU functionality does not alter any existing methods.

The best option might be to create a completely new package that offers both.