@cognusion 2015-03-03
Thanks for getting back to me. 

If a separate "New" function, ala "NewWithLRU" is used instead, there would be no API change. If that's sufficient, I'll update the pull, if not I'll maintain the fork.