@cognusion 2015-03-08
If you're interested, I significantly changed how the LRU is handled. The janitor routine deals with everything asynchronously now, reducing list walks and whatnot, and significantly improving performance when the cache is very large and under pressure. Even with extremely large caches there is no measurable performance impact vs. the pre-LRU code.