@pmylund 2015-03-09
Very interesting, thanks. Yeah, a new New func would be sufficient as long as it doesn't meaningfully impact anything else.

I'll do a little testing as well this week, but I think it makes sense to merge this.