@cognusion 2015-03-09
`.New()` has been reverted back to its previous 2-argument condition, `.NewWithLRU()` has been added with the 3-argument set. I also made some suggested README changes in a separate, subsequent commit. Feel free to ignore or edit it as you see fit.