@patrickmn 2017-04-17
I know it's been a long time. I haven't forgotten this, I'm just concerned about the overhead of Atime/Ctime when you are caching many small items and are not using the LRU functionality -- but I can't think of a good solution that doesn't involve copy-pasting almost everything.

It's really hard to gauge what the typical use case is for `go-cache`, so it's hard to know if this will explode memory usage for no reason or add a useful feature at negligible cost.

Beside this, just a few minor things (Matt, not asking you to do this, I can change this if/when it gets merged):
  - Are Ctime/Created actually necessary?
  - The NewFrom change technically breaks the API. It should probably be NewFromWithLRU (even if it's a bit silly)