@cognusion 2017-04-18
@patrickmn I'll re-review my submissions (while I've used it a ton, I haven't looked at this code in an eon) fix the NewFrom, and see what I can come up with for lightening the object size for non-LRU users (who wouldn't want an LRU?! :) )