@patrickmn 2017-04-18
Could you give me write access to your repo? I'd like to push a few changes to bring this up to current master. There's been some significant, but internal, changes recently.