@patrickmn 2017-04-19
Ok, I pushed a few commits with some changes, chiefly:

  - Changed time checks to use int64 (unixnano) since that's what the cache uses now
  - Inlined most expensive operations, only performed time.Now() once, etc. (This makes the code look pretty silly, but is unfortunately required to keep op overhead in the tens-of-nanoseconds range.)
  - Changed DeleteLRU to DeleteLRUAmount, and made DeleteLRU not require an argument (just calls DeleteLRUAmount with the overage)
  - Added a note that LRU functionality makes Get a slower, write-locked operation
  - Added a test
  - Some minor style changes

Benchmarks seem OK:

  - Before LRU changes:
     BenchmarkCacheGetExpiring-4 -                       	57.1 ns/op
     BenchmarkCacheGetNotExpiring-4 -                    	21.2 ns/op

  - After LRU changes:
    BenchmarkCacheGetExpiring-4 -                       	58.8 ns/op
    BenchmarkCacheGetNotExpiring-4 -                    	21.9 ns/op
    BenchmarkCacheWithLRUGetExpiring-4 -                	136 ns/op
    BenchmarkCacheWithLRUGetNotExpiring-4 -             	101 ns/op

The non-LRU cache has virtually identical performance, while the LRU-enabled cache has slower read operations, as expected.

Do these changes look good to you?

I am a little concerned that the unit test isn't really testing the LRU properly. Should probably have a real one...