@patrickmn 2017-05-01
Is the issue that the LRU cleanup doesn't evict the oldest items, just some old items (depending on scan/ring buffer fill order)?