@aldas 2019-05-12
this feature would be really helpful.