@cognusion 2019-05-16
In short, if you needs a true LRU this isn't for you. For large (quantity) caches, it will purge "some of of the least-recently used items" but not necessarilly **the** least-recently used items.

For my use with large caches this is perfectly fine, although I refer to it as an "Eventually Least-Recently Used Cache" in jest. I don't plan on working further on it, so whether @patrickmn wants to merge it or not is his call. It does work very very well for large caches where it's ok that the LRU math isn't spot on. I've looked into a couple different ways of making this more correct, but they result in less read performance, and longer pauses for purges: Neither of which I've been willing to saddle on this feature.